# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for RiceHub website.",
        "title": "RiceHub API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/",
    "paths": {
        "/admin/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get service-wide statistics (admin only)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ServiceStatisticsDTO"
                        }
                    },
                    "403": {
                        "description": "Admin access required",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    }
                }
            }
        },
...
        "/profiles/{username}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profiles"
                ],
                "summary": "Get a user profile with their rices by username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username (alphanumeric)",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns user (UserDTO) and rices ([]PartialRiceDTO)",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid username",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    }
                }
            }
        },
...
        "/rices/{id}/dotfiles": {
            "get": {
                "description": "Returns the dotfiles archive as a file attachment. Free dotfiles are public; paid dotfiles require ownership",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "rices"
                ],
                "summary": "Download dotfiles for a rice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rice ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File attachment"
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    },
                    "404": {
                        "description": "Rice not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rices"
                ],
                "summary": "Replace the dotfiles archive for a rice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rice ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Dotfiles archive",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RiceDotfilesDTO"
                        }
                    },
                    "400": {
                        "description": "Missing file",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/rices/{id}/dotfiles/price": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "rices"
                ],
                "summary": "Update the dotfiles price",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rice ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New price",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateDotfilesPriceDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/rices/{id}/dotfiles/type": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "rices"
                ],
                "summary": "Update the dotfiles type (free or one-time)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rice ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New type",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateDotfilesTypeDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/rices/{id}/purchase": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rices"
                ],
                "summary": "Purchase dotfiles for a rice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rice ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns checkoutUrl (string)",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Free dotfiles cannot be purchased",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    },
                    "409": {
                        "description": "Already owned",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.DotfilesType": {
            "type": "string",
            "enum": [
                "free",
                "one-time"
            ],
            "x-enum-varnames": [
                "Free",
                "OneTime"
            ]
        },
        "models.ErrorDTO": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.RiceDotfilesDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "type": {
                    "$ref": "#/definitions/models.DotfilesType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.ServiceStatisticsDTO": {
            "type": "object",
            "properties": {
                "comment24hCount": {
                    "type": "integer"
                },
                "commentCount": {
                    "type": "integer"
                },
                "openReportCount": {
                    "type": "integer"
                },
                "reportCount": {
                    "type": "integer"
                },
                "rice24hCount": {
                    "type": "integer"
                },
                "riceCount": {
                    "type": "integer"
                },
                "user24hCount": {
                    "type": "integer"
                },
                "userCount": {
                    "type": "integer"
                }
            }
        },
...
        "models.UpdateDotfilesPriceDTO": {
            "type": "object",
            "required": [
                "newPrice"
            ],
            "properties": {
                "newPrice": {
                    "type": "number"
                }
            }
        },
        "models.UpdateDotfilesTypeDTO": {
            "type": "object",
            "required": [
                "newType"
            ],
            "properties": {
                "newType": {
                    "enum": [
                        "free",
                        "one-time"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.DotfilesType"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}