# internal/handlers/profiles.go
package handlers

import (
	"net/http"

	"github.com/ricehub-io/api/internal/errs"
	"github.com/ricehub-io/api/internal/services"

	"github.com/gin-gonic/gin"
)

type ProfileHandler struct {
	svc *services.ProfileService
}

func NewProfileHandler(svc *services.ProfileService) *ProfileHandler {
	return &ProfileHandler{svc}
}

type profilesPath struct {
	Username string `uri:"username" binding:"required,alphanum"`
}

// @Summary Get a user profile with their rices by username
// @Tags profiles
// @Produce json
// @Param username path string true "Username (alphanumeric)"
// @Success 200 {object} object "Returns user (UserDTO) and rices ([]PartialRiceDTO)"
// @Failure 400 {object} models.ErrorDTO "Invalid username"
// @Failure 404 {object} models.ErrorDTO "User not found"
// @Router /profiles/{username} [get]
func (h *ProfileHandler) GetProfileByUsername(c *gin.Context) {
	var path profilesPath
	if err := c.ShouldBindUri(&path); err != nil {
		c.Error(errs.UserError(
			"Invalid username path parameter. It must be an alphanumeric string.",
			http.StatusBadRequest,
		))
		return
	}

	callerID := GetUserIDFromRequest(c)
	res, err := h.svc.GetProfileByUsername(c.Request.Context(), path.Username, callerID)
	if err != nil {
		c.Error(err)
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"user":  res.User.ToDTO(),
		"rices": res.Rices.ToDTO(),
	})
}